utilities/src/main/java/org/raml/utilities/types/Cast.java
/*
 * Copyright 2013-2017 (c) MuleSoft, Inc.
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *   http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing,
 * software distributed under the License is distributed on an
 * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND,
 * either express or implied. See the License for the specific
 * language governing permissions and limitations under the License.
 */
package org.raml.utilities.types;

import java.lang.reflect.ParameterizedType;
import java.lang.reflect.Type;
import java.lang.reflect.TypeVariable;

/**
 * Created by jpbelang on 2017-06-01.
 */
public class Cast {

  public static Class toClass(Type type) {

    if (type instanceof ParameterizedType) {
      ParameterizedType pt = (ParameterizedType) type;
      return (Class) pt.getRawType();
    }

    if (type instanceof TypeVariable) {

      throw new IllegalArgumentException("trying to get annotations from type declaration " + type + " from declaration "
          + ((TypeVariable) type).getGenericDeclaration());
    }

    return (Class) type;
  }
}
